Testing Object Movement and Spawning
Created a object spawner script
It will spawn prefabs in the positions that they are assigned, EG. background, foreground, middleground.

Created a Movement script
Assigns each object a type, background, foreground, middleground
You can set it in the editor for each object of prefab
The idea behind it is the speed of the object will change depending on the type you choose.
Only testing foreground currently.